MXS-4612 Fix alter table editor
Can not apply the changes as the schema identifier is quoted unnecessarily.
For example:

```
ALTER TABLE
   ```one_to_one```.`capital`
   CHANGE COLUMN `capital_id` `capital_id` INT(11) NOT NULL;
```

but it should have been
```
ALTER TABLE
   `one_to_one`.`capital`
   CHANGE COLUMN `capital_id` `capital_id` INT(11) NOT NULL;
```
It happens because `node.qualified_name` was used for getting schema
name which is inaccurate because the identifier is already quoted.
It should have been changed to use `parentNameData` property to get
unquoted schema name.