MXS-4545 Detect the cardinality type of the referenced table
If the referenced column is a PK or an unique index, that means
the cardinality type of the referenced table is `1`. If not, it
can only be a many type. i.e. `N`.

This also renames `index_col_names` and `referenced_index_col_names`
to `index_cols`, `referenced_index_cols`, respectively. Because each of
those properties is an array of objects containing parsed data of
index columns.