Reuse getaddrinfo results for proxy protocol header
The proxy protocol header generation requires that the remote server's
IP address is known. Currently, this information is extracted with
getpeername but this requires that the connection is fully established
before the header can be generated.

The peer address information can also be extracted by using
getaddrinfo. Luckily, we already do a call to it whenever an outbound
connection is being created. This means that if the result is stored in
the DCB (for client DCBs it already was), the same result can be reused
for the proxy protocol generation. This avoids one system call that
would otherwise be done for each new backend connection. It also allows
the header to be written in the set_dcb() function without having to
wait for the connection to fully establish. With this change, all
write_ready() functions can be made identical and could be done
automatically at the lower level.