Remove write_ready() API function
With the change in how the proxy protocol header information is
generated, there is no longer a need for the write_ready() function.

If needed, a separate API function that's called once whenever an
outbound socket has been established can be added. This approach would
be clearer compared to the dual purpose that the write_ready() function
served.