Simplify error handling in the protocol modules
The hangup() and error() API functions were nearly always treated
identically and in the cases where it wasn't, the behavior in the
error() function was lacking and did not do what the hangup() code did.

Given that the lower-level code knows the exact reason why the error
happened or is at least better equipped to deduce a possible reason, it
is thus better to provide an error description into the new error()
endpoint instead of forcing the protocols to deduce it by themselves.

If a EPOLLERR event is received, the lower-level DCB code now extracts
the socket error if one is available and passes that as a part of the
generated error message. This should help identify the true source of
network errors better while also allowing the protocol code to do less
work.

There was also a static hangup() function in the DCB that was generated
the fake hangup events. The function was renamed to generate_hangup() to
better describe its purpose. It also now takes an error message which
helps identify the source of the fake event. Both the act of forcing a
server into maintenance and the fake hangups generated by monitors will
now clearly show in the error messages which will reduce some of the
confusion that the old API generated when it came to hangup events.