MXS-4590: Optimize marker generation
The markers are now pushed onto the marker stack without checking for
bounds. This reduces the cost of pushing more than one marker by
counting how many bits are set and allocating enough space to hold them
all with one call to resize().

The make_ascii_bitmap() function now returns a std::array. This allows
the values to be converted into global values which remove the need to
initialize them that in turn gets rid of the guard variables that would
otherwise end up being checked every time. In the grand scheme of things
this is unlikely to be of any significant importance.