MXS-4590: Make the markers an implementation detail
Since the markers are only used internally by the query canonicalization
and the multi-statement detection, they should be hidden from the
API. This both makes it simpler to use while at the same time it allows
the multiple redundant thread-local Markers objects to be eliminated.

In practice the markers were duplicated roughly three times: one in
pp_sqlite for get_canonical(), one in QueryClassifier for
is_multi_stmt() and once again in QCInfoCacheScope for
get_canonical(). In the worst-case scenario, the markers could be
duplicated nearly six times as some filters also used direct calls into
the maxsimd functions.