MXS-4590: Use 32-bit offsets instead of pointers
The SQL strings that are being parsed are always less than 4GB
long. This means that we can use a 32-bit offset into the string instead
of a 64-bit pointer to store the marker generation result. This cuts
down the marker memory usage in half.

In practice, the maximum length of a query that MaxScale will parse is
16MB long. Since each character can theoretically produce a marker, this
means that previously up to 128MiB of memory could be taken up by the
markers which is now reduced to 64MiB.