2.5.26 Fix assert failure in runtime_create_volatile_server()
That assert was the result of the commit with UnmaskPasswords
having been cherry-picked into 2.5. UnmaskPasswords is not
thread-safe, so the assert was correct as such. However, the
Xpand-monitor calls this function directly and thus the assert
will always be triggered.

This is now fixed by adding to UnmaskPasswords a static recursive
mutex that is locked in the constructor and unlocked in the
destructor, which allows the assert to be removed.