MXS-4620: Optimize DCB::read() usage for RVO
The tuple that the function returned was not returned as-is and a
temporary tuple ended up being constructed multiple times for each read
from the client. By using structure binding, references to the elements
can be used instead.

Since epoll reports an EPOLLOUT event whenever the socket is writable,
the DCB::writeq_drain() function is always called before each read. This
means that the buffer recycling code as well as the buffer clearing code
was called even when the buffer was empty. Checking that a change in the
emptiness of the buffer happens will avoid the checks completely.