Remove dead code related to mxs::Buffer
Removed the mxs_response_to_string function that appears to be
unused. Since the mxs::Reply class already provides this information on
the info level, the utility of this function is questionable.

Removed references to mxs::Buffer in the resultset generation code that
was missed in the commit that changed them to return a GWBUF value.

Removed the unused using declaration of mxs::Buffer::iterator in the
RWBackend code as well as the unused mxs::Buffer variable.