MXS-4619 Build pp_mysqlembedded using server build
From 10.5 onwards, the embedded library is no longer included in
the MariaDB server package. So the server will have to be built
if pp_mysqlembedded is to be built.

This change now only builds pp_mysqlembedded using a 10.4 server
build instead of a 10.4 server package. Subsequent changes will
then take a newer server version into use.

With this change, anything older than 10.4 can also no longer
be used.

FindMySQL.cmake is removed as the only component that needed it
was pp_mysqlembedded, something which it no longer does.