MXS-3735: Add input to change ERD worksheet name
Prior to creating the ERD worksheet, users need to have the ability
to assign a name to it.

This also disables the `enter` key event for submitting the form.
When using accessibility mode, users rely on the tab key to move between
various HTML elements. The enter key is specifically designated for
selecting checkboxes in the `selectable-schema-table-tree` component and
cannot be utilized for any other function.