MXS-4622 Facilitate the usage of `entity-editor-ctr`
The height percentage of the graph is now stored per ERD task. This
ensures the height of the graph is adjustable differently between
ERD worksheets.

This adds double click event on each node. Clicking on a node will
now show `entity-editor-ctr` component, the graph height will reduce
from 100% to 50%.

This also replaces the watcher of graph data with node keys to prevent
the graph from being rerendered unexpectedly when node attributes are
changed.