MXS-4549: Remove close reasons from mxs::Backend
The information was largely redundant or useless and the context of the
error is either evident from the error itself or was also logged
alongside the verbose status.

The verbose status that's used in readwritesplit can be created when
needed based on information that's available elsewhere. The only case
where it's really useful is when routing of queries fails and the status
of the servers needs to be known.

The error logged by readwritesplit is also now on a single line. This
makes the log parsing that's done by the REST-API work with these error
messages.