Remove call to session_ok_to_route()
The function shows up in profiling and the checks that it does should
not be needed as they would usually indicate an error
condition. Simplifying the check to only inspect the session state has
the same end result. The debug assertion inside the branch that calls
session_ok_to_route() should make sure the logic remains the same.