MXS-4622 Persist graph layout
If the graph has been laid out or nodes are repositioned by the users,
`mxs-erd` component should not run the simulation again. In other words,
the coordination of nodes should be persisted.

This adds `is_laid_out` field to `ErdTask` model and persists
node coordinates.

The also refactors the way the graph being rerendered when generate
new ERD for an existing ERD worksheet by relying on the id of nodes.