MXS-4549: Store statement level checksums in Trx
Storing a checksum per statement makes it possible to detect which
statement caused the checksum mismatch. Any statements that are ignored
due to transaction_replay_checksum are added as an empty checksum. This
allows the checksum code to always assume that a checksum is added even
if all statements would be ignored.

The checksum algorithm was also changed from SHA1 to the CRC32 function
provided by zlib. The checksum is needed to verify that the transaction
is executed with the same result. SHA1 was used previously as it offered
better collision avoidance but now that there is a checksum per
statement, CRC32 should be adequate for most cases. The checksum
algorithms could be made configurable but an evaluation of the relevance
of this for transaction replay must be first done.

The execution of the interrupted statement now also has its own
state. This is needed for result splicing as the code that will do the
splicing must only be called when the transaction has been successfully
replayed and the interrupted query is being retried.