MXS-4549: Replay partially received queries
A partially received query is one that the server started to send and
which has been partially processed by readwritesplit but the trailing
end of it was lost due to a network error. A good example of this is a
large resultset that gets cut in half due to a network outage or a
stored procedure call that fails halfway through its execution.

Stored procedures are especially prone to fail mid-result: the server
starts sending results as soon as they have finished and there might be
an arbitrarily long delay between the start of the result and the final
part of it.

Queries that end up being partially received are now replayed and the
results are spliced together to look like one contiguous
result. Currently this is only done for queries that were interrupted
while a transaction was open. The feature could be extended to cover all
queries but it would require result checksums to be always calculated.

This works on the assumption that a byte offset and the checksum will
avoid the need to track packets or individual results. The protocol
modules will take care of not delivering partial packets which prevents
packets from being cut in half.