MXS-3735 Establish two-way binding for zoom and pan values
The zoom and pan values will be later changed in `mxs-erd` to auto fit
the nodes within view.

This also sets scale extent to [0.25,2] to limit the scale. The zooming
can scale the elements from quarter their original size (0.2.5) to
double their original size (2) as there no need to scale smaller or
larger than that.