MXS-4232 Store previous pwd in BaseUserManager and use it
When the password of the service user is changed, the previous
password is retained. If the updating of the users fail with
an authentication error, then the updating is retried using the
previous password. The purpose of this is to remove the need for
issuing the following two commands simultaneously

$ maxctrl alter service MyService password=newpassword

MariaDB [(none)]> set password for <the_service_user> = password('newpassword');

The necessary changes in the user managers are provided in subsequent
commits.