MXS-4232 Store previous pwd in BaseUserManager
When the password of the service user is changed, the previous
password is retained. When connecting, derived user managers
are expected to try the previous password if the current one
does not work. The purpose of is to remove the need for
issuing the following two commands simultaneously

$ maxctrl alter service MyService password=newpassword

MariaDB [(none)]> set password for <the_service_user> = password('newpassword');

The necessary changes in the user managers are provided in subsequent
commits.