MXS-3735 Add zoom control dropdown
The dropdown has several options to zoom the graph. There is a special
option which is called `Fit` that fits the graph into the view.

The `panAndZoom` object is now moved to `diagram-ctr` component.

This also beautifies the look of the `er-toolbar-ctr` by increasing the
height to 40px, reordering certain icons, adding margin between icons
and adding divider element.