MXS-4549: Implement mxs::Checksum using CRTP
The classes that inherit from mxs::Checksum were never used via a
reference to the base class. They were, however, used by-value and were
compared and copied around. This makes accessing them via base class
pointers somewhat complex as comparisons and copying would need to be
implemented. Using CRTP makes it possible to move the overloads into the
base class and only implement a handful of functions in the derived
class.

Moved the implementations into maxbase as they no longer operate on any
MaxScale-specific data types. and renamed the algorithms to be less
verbose and to avoid the naming conflict that arise e.g. with SHA1 being
a function name and a class name.

Because the zlib.h header is no longer included by <maxscale/utils.hh>,
a few fixes were needed elsewhere.