MXS-4232 Add alternate_password to mxq::MariaDB
If connecting fails due to ER_ACCESS_DENIED_ERROR and alternate_password
has been set in the settings, a connection attempt using alternate_password
will be made.

This is to ease the changing of passwords, as that in principle means
that the password in the MariaDB server and the password in MaxScale
must be changed simultaneously. With this change it is straightforward
to arrange that the MaxScale password can be changed first, and the
MariaDB later.