MXS-4622: Move `graph_height_pct` and `active_entity_id` to temp model
Keeping `active_entity_id` to show DDL editor is a nice to have feature
but that adds extra complexity to handle case when there is no connection
bound to the ERD worksheet. To make it simple, the value shouldn't
be persisted.

By moving `active_entity_id` field to `ErdTaskTmp`, the `graph_height_pct`
field should also be moved.