MXS-4549: Add mxb::xxHash
The included code is a pruned down version of xxHash 0.8.1. All of the
support files, documentation and other auxiliary files were removed. The
library can be made into a header-only one with the XXH_INLINE_ALL
define. This simplifies its usage as the functions can be declared
inline and the library types can be used. The library appears to be very
clean in the sense that it does not expose any conflicting constants and
has a mechanism for prefixing all functions and values if this turns out
to be a problem.

The xxHash class uses the 128-bit XXH3 hash function from the
library. This should be a good alternative to the existing SHA1 and
CRC32 checksums when it comes to checksumming resultsets in
readwritesplit. A micro-benchmark comparing against SHA1 and CRC32 seem
to suggest that xxHash is almost an order of magnitude faster in the 8
to 32kb range. This seems to be in line with the benchmarks done on the
hashing algorithm and it verifies that the build is working correctly.

Replaced the CRC32 checksum with the xxHash checksum in readwritesplit
and added unit tests for the new class.