MXS-4627: Make SessionStats usage thread-safe
The values are now accessed correctly via the WorkerLocal instead of the
reference.

In readwritesplit, the inc_total() call was done only when the query was
successfully routed whereas the read and write increments were always
done even if the query failed at a later stage. In addition, the
statistics for the router level reads and writes and the session level
reads and writes used a different logic. Given that the TARGET_MASTER
and TARGET_SLAVE types are intended to be an abstraction of reads and
writes, it's better to use this one value instead of trying to deduce it
from the query type mask and server status.