MXS-4628: Let connections die naturally
Pinging the connection makes wait_timeout and interactive_timeout
ineffective. A better method is to peek into the TCP socket and see if
it fails. This does not affect the idleness of the connection on the
server side. It also gets rid of the possibility of unwanted
reconnection happening as a result of the ping.