MXS-4123: Improve mxs3663_universal_causal_reads
The test now takes into account the inability to replay a transaction by
comparing the maximum value stored in the table instead of the count of
values in it.

Adjusted the mxs4123_fast_universal test to be more tolerant of replay
failures. The act of writing on the master node seems to cause locking
conflicts but turning on transaction_replay_retry_on_mismatch seems to
solve them.