Fix a minor UX issue with preview table data or details
If the users alter a table using the DDL editor, that table node
is highlight as an active node in the schema tree. After that, if
the users click on other table or right-click and choose
`Preview Data (top 1000)` or `View Details`, the active node changes.
This action shouldn't be allowed in the DDL editor.

Clicking on a table/view node to preview the data is a confusing UX.
The event actually does two things: Expand the node and send a query
to either select top 1000 or describe table. If the users just want
to expand the node, this behavior is annoyed.

This adds a "view table" icon and disables preview options when
there is an altering node. By doing this, the `activatable`
property of the node is no longer needed and has been removed.