Create ParameterStatus packets for clients
Certain parameters are always sent on startup and clients like psycopg2
expect them to be sent. MaxScale currently sends none and at least the
documented set of parameters should be sent. Most of the information
needed to send them is already available but some values, like the
server_encoding and DateStyle, need to be probed from the actual
servers.