Don't include <maxscale/utils.hh> unless needed
The file was included by <maxscale/pcre2.hh> that in turn was included
by <maxscale/config2.hh>. The latter is included nearly always by all
code in MaxScale and thus <maxscale/utils.hh> was also included.

Removed the Closer class and replaced its use in the masking filter with
explicit calls to the functions it used. The class was not really needed
as the value was either passed to another object or could be immediately
closed after being used. None of the code that is called throws
exceptions which makes this a safe thing to do.

Moved the Registry template into its own file since it's included by
<maxscale/session.hh> which is included by most of the MaxScale code.