MXS-4632 Add Parser::relates_to_previous()
Returns true if the statement contains something that implicitly
makes it relate to the previous statement and hence need to be
sent wherever that was sent. In the MariaDB case use of the function
FOUND_ROWS() would make it related to the previous.

Currently this is handled by the query classifier that asks for all
functions and then looks for FOUND_ROWS(), but that is not robust as
for instance Postgres lacks that function but may have something else
that makes a statement related to the previous one.