MXS-4622 Turn `ddl-editor` into a reusable UI component - part final
Move the handlers of `on-revert` and `on-apply` events to a newly
created component called `alter-table-editor` which is used for
the Query Editor.

The `mxs-ddl-editor` is now refactored to a "presentational"
component that receives data via props and emits events to
handle revert and apply buttons.

This also removes obsolete codes in subcomponents of `mxs-ddl-editor`.
Subcomponents shouldn't have `activated` and `deactivated` hooks or having
manually watcher because that makes it harder to reuse the `mxs-ddl-editor`
component. Those hooks and watchers can now be removed because the ddl
component is now rendered only when it has data.