Split version.hh.in into two parts
Splitting the version information header into an actual version part and
the build information section makes it possible to use ccache and other
compiler caches to cache the objects. If the commit hash is included as
a compile-time constant, the object files from two different commits
would never match. Splitting the file in two allows the version number
to be used as a compile-time constant while still making it possible to
isolate the commit hash into just misc.cc.