MXS-4622 Make `mxs-ddl-editor` compatible with the output of TableParser - part 1
This renames some properties in `tbl_creation_info` and unquote
identifiers in the parser.

This enhancement further improves the functionality of the `unquoteIdentifier`
helper by enabling it to identify different types of quotation marks.
For instance, when handling the value of the comment field, which is
consistently enclosed in single quotes.