MXS-4618: Make ExternalCmd execution asynchronous
The new start(), wait() and try_wait() methods are used to implement the
old externcmd_execute() function that is now renamed to run(). The
functionality is the same for the run() command which is implemented
using start() and wait(). This allows the command execution to be done
in steps which is a prerequisite to being able to pipe data into the
command.

Added unit tests for the wait() and try_wait() functionality.