MXS-4618: Use X3 for parsing SQL
Now that the boost headers are reusable, the parser can be implemented
using Spirit X3. The validity and parsing of the S3 URL is also done at
the same time which makes its use easier.

Currently the remaining unparsed SQL is treated as always valid. This
means that in the case where the data is fed into Xpand, the error is
noticed too late and the client connection will be closed when
MXS_SESSION::kill() gets called.

Added a unit test that covers some of the basic use cases. More cases
can be added if/when the parsing is extended.