Remove unused parts of <maxbase/random.hh>
The StdTwisterRandom class was only used in test_utils.cc and it was the
only reason why <random> was included. The header is pretty big and if
it's not needed it's better to remove it to speed up compilation.

Changed the implementation of XorShiftRandom::rand_bool() to use modulo
2 for boolean values. Given that the return vale of rand() is a powers
of two, there should be no bias to the randomness.