Make rwsplit backend selection smarter
The best_score function did a lot of extra work which can be avoided by
making sure the backend selection functions are only called when an
actual selection must be done. With less than two candidates there's no
need to inspect the relevant metrics.

The conversion into a double was also unnecessary. This was originally
done to make server weights work but they no longer exist which means
the comparisons can use native types. The last usage count also
unnecessarily calculated the true idle time when all we need to do was
to compare the timestamps. This gets rid of the clock calls in the
function.

Converting the metric selecting function into a template gets rid of the
std::function construction. This appears to result in the same outcome
as using `template<auto server_score>` would except that this approach
needs a smaller set of changes. This is what makes the native type
comparisons possible.