MXS-3664 Change ReplyRoute from std::vector<Endpoint*> to class
Now efficient to add more endpoints when moving up the chain,
without a need for casting the constness away. Further, each
ReplyRoute contains a pointer to the first Endpoint, so that
there never is a need to iterate to find it (not that the
route would ever be deeper than a few levels at most).