MXS-4622 Handle side-effect of changing keys
Keys of a table are parsed into a `keys` object. However, the
`mxs-ddl-editor` don't modify the value of each key directly but
rather using 2d array structure. The toggle action for PK or UQ
attribute will return a boolean value, so it's not the same datatype
as the value in the `keys` object and must be handled as a side-effect
action. In other words, the table editor in `mxs-ddl-editor` reads and
updates values in `definitions.cols` 2d array, if PK or UQ values
are changed, `definitions.keys` must be updated.
```
definitions: {
    cols: [[]], // 2d array
    keys,
}
```
The right approach is to keep the parsed columns as an array of objects
structure, but because the DDL editor relies heavily on the
`mxs-virtual-scroll-tbl` component, the structure was changed to
2d array, so the side-effect is unavoidable.