MXS-3664 Add function for releasing a response
A filter that is capable of terminating the request processing
and deliver a response immediately, can explicitly set the response,
which then will be delivered by the session.

When CacheFilterSession is used in NoSQL protocol, this has to be
handled manually as CacheFilterSession is before the session in
the request processing chain. Thus, functions for detecting whether
a response has been set and to release the response are needed.