MXS-4622 Add setting button for each entity
Double-clicking on the entity to alter table is not an intuitive
approach as it's not the most obvious or discoverable method.
So, in addition to the double-click event, this adds setting icon
button to show entity option menu.

With this menu, the users can choose the following options:
Alter Table, Drop Table.