MXS-4622 Remove the watcher for node ids in `entity-diagram`
Having the watcher there helps to redraw the diagram but if a node
is added or removed, it would also trigger the redraw. The `init`
method should be called on the first render of the graph or
when the users generate new ERD for existing ERD worksheet.
Adding or removing a node should not show the loading indicator
as it's a low performance action.

This removes that watcher and adds a `key` property to `ErdTaskTmp`
model. This key can be used for re-render the diagram. The `key`
can be directly assigned to `entity-diagram` component to
trigger re-render the component but that can have performance
implications as multiple ERD worksheets will be rerendered at once.
Therefore, using a watcher to assign the key to component'state
is much more efficient.