MXS-4631 Move rpl_state to class BinglogIndexUpdater
The rpl_state is strictly speaking only needed for
diagnostics and "show slave status".

This seemingly complicated manner to keep it accessible
via config is so that the Reader (the side that
reads files and sends events to replicas) can access
it on thread-start and wait for the rpl_state to catch
up with the rpl_state the replica requested, rather
than cause the replicas to fail and give up.
Especially useful when testing, but also when a new
master is being restored from backup.

The Reader was written to be able to run as a separate
process, on a different machine than the Writer. This
is now partly broken, but all that is needed is to
make the way it finds the latest gtid configurable.
Commented in reader.cc. Alternatively, the Reader could
simply reply with an error if the requested gtid does not
(yet) exist, like the master does.