MXS-4215: Allow manual invalidation of cached entries
Compared to invalidation, the clearing of the cache is an aggressive
operation that causes all clients to block. With the introduction of
`max_staleness`, the invalidation of entries should be preferred over
clearing them. Using invalidation to implement scheduled updates is
better in the case when there's always some traffic on the system.