MXS-3664 Introduce internal cache in NoSQL protocol
When an internal cache has been enabled, a CacheFilter instance
is created inside nosqlprotocol. That makes it easy to make all
cache configuration options available as such, without any extra
work.

Then, for each ClientConnection instance (i.e. client) a
CacheFilterSession is created. There are a couple of problems
with that:
* A filter is a mxs::Routable while a client connection deals with
  mxs::Components.
* The CacheFilterSession sits before the session in the routing
  chain, so packets will not automatically be delivered to it.

This is dealt with by introducing two classes, one for making
an mxs::Component appears as an mxs::Routable, one for making
an mxs::Routable appear as an mxs::Component.

The end result is that the entire machinery is handled by
ClientConnection and the rest of nosqprotocol can be oblivious
to the arrangement.

This does not yet provide any advantages compared to just using
the cache filter in a service, but in the following steps, the
NoSQL protocol packets will be used both as keys and values, which
will remove the BSON <-> MariaDB protocol packet translation.