MXS-4629 Separating initial and staging graph data
When a new node is added and created in the database, detecting
the editor mode for that node becomes a hassle task. Having two
separate objects would make it a lot easier to get the diff.
The initial object is kept in `ErdTask` model while the
staging object is in `ErdTaskTmp`. The initial object keeps
the final version of the ERD. i.e. When a table is succesfully
altered or created.

The `mxs-ddl-editor` will be later extended to introduce
a button to save changes instead of executing the queries
directly to the connected database.

The users have two interactive approaches in the ERD worksheet:
 - Alter/Create a table -> Click the "Apply" button to execute the
queries directly to the connected database.
 - Alter/Create a table -> Click the "Save" button to save changes
to staging object. There will be a button in the toolbar to execute
all changes at once or output queries.