Add missing ops to xrouter
The operations that were added to the parser weren't added to the
xrouter. Changed the code so that any new op additions must be
explicitly added to the xrouter as well. This should help avoid unwanted
changes in behavior when the parser is extended but the xrouter isn't
handling that case yet.