MXS-4629 Handle show node options
If an entity is in staging data but not in initial data, it means
the "CREATE TABLE" query for it hasn't been executed. Hence, it
needs to have different labels for node options. The labels
will be `Edit` and `Delete` for that entity while the existing
entity will have `Alter` and `Drop` options

This also moves the setting button to the parent component
of `entity-diagram`.