MXS-4629 Show `schema` input in `mxs-ddl-editor`
When creating a new table, the schema name must be defined.

This uses `v-combobox` to show all schemas available in the ERD so
that users can quickly select one or manually type a new schema name.

This also removes `mode` string props as it's unnecessary. A simple boolean props
is enough.