Fix XgresSession password parsing
The operation was also wrong here. In addition, the offset into the SQL
string was wrong because the `pw` value was calculated from a
substring. It should've included the `role` value as well. Now that the
parser is able to identify the `CREATE USER` and `CREATE ROLE`
statements, there's no need to find those tokens anymore and the problem
of the wrong offset goes away when the search for those tokens is
removed.