Add versions/VERSIONDEVELOP.cmake
The version of a package built in `develop` is now 33.02 instead
of that of the previously released version. 33.02 is 10 years
into the future, so it will work for a while.

So, at release branch branching time, a VERSIONXXYY.cmake file
has to be created and VERSION.cmake updated. That's ok as you
anyway need to remember to update MAXSCALE_MATURITY.

An alternative would have been to create at release branch branching
time the next VERSIONXXYY.cmake in develop. However, that would mean
that 1) you ought to be pretty certain what the next version will be
and 2) that a package built in develop would look suspiciously like
a "real" package (although, slightly better that the current situation
where it is indistinguishable from the last "real" package).

The maturity of 'develop' is always "Develop".