MXS-3735 Make `execute-sql-dialog` editor readonly
Allow the users to freely modify the auto-generated script would
cause unforeseen behaviors or bugs. e.g. If the users modifies
the genereated script for altering a table. If would be difficult
to parse any new modifications makd to the script and assign the
data back to the editor.

The dialog will be later extended to have an option for only exporting
script. With that option, the users can modify and save the script.