MXS-3735 Handle delete links of a node
After deleting a node, its links should also be deleted.
This also removes `DROP` option as it brings complexity to
ERD and it is also not a usual action. The users uses
ERD mainly for generating ERD, adding new tables, modifying
and add FKs. However, the users may delete the table entity
from the diagram but not from the database.