Make mxb::Cipher thread-safe
Changed the way that the EVP_CIPHER_CTX is used. Since the OpenSSL APIs
do not have a consistent way to reuse it, a new context is allocated
each time something is encrypted or decrypted. As the EVP_CIPHER itself
is const, it can be used concurrently by all threads.