MXS-4629 Remove the `Save changes` button
In hindsight, having that button is unnecessary and inconvenient.
The changes in the ddl editor can be saved to `ErdTaskTmp` model
immediately. However, there could be a slightly laggy behavior
while the users typing text in the input, but that's mostly
neglectable in production build. Anyway, debounce or throttle
for text inputs can be later introduced to handle that issue.