MXS-3664 Prepare for both MariaDB and NoSQL cache values
If a NoSQL response is cacheable, it will be cached using
a key created from the NoSQL response, with a specific tag
added. That way it can be distinguished whether a cache value
is a MariaDB response (i.e. a result set) or a NoSQL response
that can be used as such.