MXS-4654: Fix data race in mxs::ListenerData usage
The std::shared_ptr m_shared_data was assigned to in
Listener::post_configure() while it was concurrently being accessed in
Listener::accept_one_dcb(). The control block of a shared_ptr is
thread-safe but the shared_ptr itself is not. This is why C++20 added a
specialization of std::atomic for shared_ptr.

This has been a problem in MaxScale for a long time but the likelihood
of an actual data race happening is very low. However, it's still
possible and the probability of it happening increases with the number
of concurrent client connections.

Fixed test_poll to not call functions on the RoutingWorker that aren't
expected to be called there.