Simplify RouteInfo
The simplification of the RouteInfo struct and its members shrinks the
size down to 16 bytes which makes the copying a lighter operation. This
is beneficial as a backup of the old state is always copied into the
QueryClassifier in case a revert_update() call occurs later.

Instead of always copying the temporary tables in case they are rolled
back, a delta can be stored instead and the revert operation can
deferred until revert_update() is called. This assumes that the creation
and deletion of temporary tables happens less often than all other
operations.

Also removed some redundant functions from RouteInfo that weren't
needed.