MXS-4635: Add connection metadata
The OK packet that's sent as a response to the authentication now always
contains a `threads_connected` system variable change that contains the
current number of connections to the given service.

More metadata can be added with the `connection_metadata` parameter that
overrides any default values that MaxScale sends (currently only
`threads_connected`).