MXS-3664 Use an SQL based key when caching Result Sets
The NoSQL response should be cached using a key generated from
the NoSQL request, but the result set should be cached using a
key generated from the SQL.

The reason is that two different NoSQL request may correspond
to the same SQL (e.g. a 'Find' where the used 'batchSize' is
different). With the result set key generated from the SQL, both
will benefit from the result set being in the cache.

Now things work the way the documentation claimed it would.