MXS-3664 Provide Command::table()
It was earlier provided in the template PacketCommand from
which all commands but OpMsgCommand derived and separately
in OpMsgCommand.

Reason is that CacheSessionFilter::invalidate_words() is not
reliable but it may return nothing by the time the result would
be needed.

The table provided in the NoSQL command can be used for the
same purpose, and that will always be available.