MXS-4216: Add test case
The test checks that the stale entries are correctly accounted
for. Actually testing that they end up being used is more complex as it
requires either a massive amount of tables in the database or an
artificial way to slow the query down which currently doesn't seem to be
possible.

The module commands that invalidate and clear the cache are also tested.