MXS-4657: Improve errors for unknown objects
The module command API endpoint can be easily improved to generate
errors for unknown commands. For unknown objects, this is a little
tricky and would require a change in how the endpoint search is done. As
such, this is the only improvement that can be done in 2.5.

Also added a clarification to the MaxCtrl error message whenever a HTTP
404 response is given. Since the endpoints themselves should always be
correct, the only thing that can go wrong is the object name or type.