MXS-4527: Don't use dots as separators
This cannot be relied on since database names are allowed to have dots
in them. The improvements to the parser that were done for 23.08 make it
possible to always use a correctly parsed table name in all cases.

Added a test case that reproduces the problem and verifies that it's
fixed by this commit.