Fix auth_wchost
The user that was used for the netmask test was created after the latest
update of the users. This resulted in an authentication failure as the
user was not found and an update could not be made as one was recently
one. To fix this, the users can be forcefully reloaded with MaxCtrl.