MXS-4629 Use column id, node id for finding relationship
The DDL editor gives the users ability to alter the table name and
column name but those are used for finding relationships between
tables. Side-effect should be avoided as much as possible.

Using generated column id and node id is a more reliable solution as
it keeps the relationship between tables even when a column or table
name is changed.