MXS-4656: Make trx replay require default trx tracking
If `session_track_trx_state` is enabled, `transaction_replay` and
`optimistic_trx` are both automatically turned off. This is done to
avoid problems that arise from the inadequate transaction state tracking
that would be caused by having them on.

Updated the documentation to reflect the current state of the
feature. With the query classifier cache, parsing the SQL is not as
expensive as it once was when this feature was first added.