MXS-4629 Extend `TableParser` to automatically generate UID
UID for table and its columns can be done in `TableParser`.
This helps to transform FKs easier without having to iterate
over parsed tables to generate nodes and transform the `cols`
definitions data structure into 2d array format.

The `tableParserTransformer` now accepts `parsedTables` parameter
which is used to transform FKs.

This also extends `TableParser` to assign `schema` in table `options`
as it helps to reduce the complexity of creating a hash of parsed
tables, keyed by schema name.